#80 refactor: UserInfoUseCase 제거하고 로직을 LoginUseCase로 이동
