+++ Projects/Features/FeatureLogin/FeatureLoginUseCase/Sources/Interface/LoginUseCase.swift
@@ -11,6 +11,8 @@ import RxSwift
 import CoreEntity
 
 public protocol LoginUseCase {
+    var userInfo: Observable<UserInfoEntity?> { get }
     func saveUserInfo(_ userInfo: UserInfoEntity) -> Completable
     func fetchUserAuthInfo(for loginType: LoginType) -> Observable<UserAuthInfoEntity?>
+    func clearUserAuthInfoIfNeeded() -> Completable
 }
